// frontend/src/app/api/model/defaultPage.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const DefaultPage = {
    Dashboard: 'DASHBOARD',
    Tasks: 'TASKS',
    Habits: 'HABITS',
    Learning: 'LEARNING',
    Plans: 'PLANS',
    Settings: 'SETTINGS',
} as const;
export type DefaultPage = typeof DefaultPage[keyof typeof DefaultPage];