// frontend/src/app/api/model/settings.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DefaultPage } from './defaultPage';


/**
 * Profile and preferences of the single user (FR-10.2, Q3)
 */
export interface Settings { 
    displayName?: string | null;
    planStartNotifications: boolean;
    defaultPage: DefaultPage;
}
export namespace Settings {
}